requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
cloudinary==1.44.2
Django==6.0.4
django-cloudinary-storage==0.3.0
djangorestframework==3.17.1
drf-yasg==1.21.15
gunicorn==25.3.0
idna==3.11
inflection==0.5.1
packaging==26.1
pillow==12.2.0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
